requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
google-api-python-client==2.141.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1